requirements.txt
Jinja2
kubernetes
PyYAML
requests
urllib3
